Move DDBClient to new client folder
